Place holder Project Web App
Hi all,
Skeleton/place holder site for the web app.
Go through it and see if it works for you.
All links are live and login/logout take you to the relevant parts of the web app.

I'm sketching out some things to put on the site (basic images, heading etc)